Initial commit of distributed fork-join
This isn't ready for consumption at this point; it's directly accessing
to JGroups, which isn't correct within an Infinispan context. But it's a
start.
Next step: decide whether to replace the JGroups bits with Infinispan's
abstracted RPC, perhaps including some custom commands, or else actually
use an auxillary "helper cache" to cache the tasks themselves. Need to
think that through...