Initial implementation of the TaskCachingDistributedForkJoinPool
This is an implementation that uses a "background cache" that mirrors
the partitioning structure of the the cache used to create the pool, and
holds tasks. The hope is that the tasks then are partitioned across the
cluster in the same way that the data is.

This implementation is incomplete; notably, there is no task failover,
a lot of DistributedExecutor methods are not implemented, and there are
still a number of TODOs in the code that need to be cleaned up. But, the
basic unit tests are passing.