A quick sketch of what a DistributedFJTask with listeners might look like.
I'm torn on this one; on the one hand, it's nice to let it notify
callers directly, and it's kind of convenient to have it effectively be
a NotifyingFuture (we could even implement that interface, actually,
although this example doesn't). 
However, it's a bit more intrusive. People extending this have to
implement doCompute() instead of compute(). That's a bit annoying.