JGroupsDistributedForkJoinPool now working
oddly enough, TaskCachingDistributedForkJoinPool is now failing tests -- the last merge from master broke the tests' clustering. I haven't checked to see if it fails in real life too; that's a little disturbing.

We're taking a short-term approach of registering a listener on the actual JGroupsTransport itself; this is hacking fairly deep into Infinispan's internals, but it works for now. I suppose if we end up moving ahead we'll want to use the Command approach that Infinispan uses in most places.